requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
cryptography==42.0.2
PyJWT==2.8.0
waitress==3.0.0
python-dotenv==1.0.0
Flask-WTF==1.1.1
python-magic==0.4.27 